Filter the templates where they are written, not where one caller writes them
Three places echo the stored template raw, each carrying a phpcs:ignore
justified by "kses'd on save" -- and that was true only of the Settings API
path. register_setting() attaches its sanitize callback on admin_init;
maybe_upgrade() runs on init at priority 1 and writes through
Options::save(), so the migration never met it. The row it folds in comes
from 1.78.1, which stored trim( $_POST[...] ) with no filtering of any kind,
so a hostile template was copied across verbatim and echoed to every visitor
by a plugin that believed it had been cleaned. WP-CLI, cron, a restored
backup and any other caller reach the same door.

The retired-keys drop was already in save() for precisely this reason, and
the docblock already said so; the templates now sit beside it.

record() -- the deferred counting path -- guarded with get_post_status()
alone, which answers with a truthy string for every row in wp_posts:
revisions, autosaves, auto-drafts, attachments, trash, drafts,
nav_menu_item, wp_block, wp_template. So it stopped ids that named nothing
and accepted every id that named anything, which is not what its docblock
claimed. An unauthenticated caller could walk the id space writing a views
row against each one, and read back the count of unpublished posts on the
way. It now asks for what the wp_head path already knows: a real, publicly
viewable post that is not a revision.

And render_item() casts the meta, the way Display already did. The key is
unprefixed, so is_protected_meta() does not cover it and any author can set
it to a string through Custom Fields -- and both count tokens are built
eagerly whether or not the template uses them, so one such post turned every
listing including it into an uncaught TypeError on PHP 8.

+++ includes/class-wp-postviews-query.php
@@ -87,7 +87,13 @@ class WP_PostViews_Query {
 	 * @return string
 	 */
 	protected static function render_item( $chars ) {
-		$post_views = get_post_meta( get_the_ID(), 'views', true );
+		// Cast, the way WP_PostViews_Display does. The meta key is unprefixed and
+		// therefore unprotected, so anyone who can edit a post can set it to a
+		// string through the Custom Fields box -- and both tokens below are built
+		// eagerly whether or not the template uses them, so on PHP 8 a single
+		// such post turns every listing that includes it into an uncaught
+		// TypeError from number_format_i18n() or from the division.
+		$post_views = (int) get_post_meta( get_the_ID(), 'views', true );
 
 		/*
 		 * Escaped here rather than left to snippet_text(). Until 2.0.0 the only
